+++ CHANGELOG.md
@@ -2,3 +2,7 @@ Alle Aenderungen von main werden hier dokumentiert:
 
 - stories.md auf STORIES.md umbenannt, weil borko vielleicht automatische
   scripts hat.
+- add vue routing
+- add list page
+- add settings page
+- fuege Abfrage von Namen im Dialog ein